requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
litellm>=0.1.0